15-23

+++ app/TDO/ShopCsvDataTdo.php
@@ -11,24 +11,24 @@ class ShopCsvDataTdo
     public ?string $kod_t;
     public ?string $sayt_tab;
     public ?string $foto;
-    public int $deb_kol_kon;
+    public float $deb_kol_kon;
     public float $tsena1;
     public ?float $tsena2 = null;
     public ?float $tsena3 = null;
 
 
-    public function __construct( array $data ){
-
+    public function __construct(array $data)
+    {
         $this->firma = $data['firma'] ?? null;
         $this->naimenovanie = $data['naimenovanie'] ?? '';
         $this->dobavka = $data['dobavka'] ?? null;
         $this->kod_t = $data['kod_t'] ?? null;
         $this->sayt_tab = $data['sayt_tab'] ?? null;
         $this->foto = $data['foto'] ?? null;
-        $this->deb_kol_kon = $data['deb_kol_kon'] ?? 0;
-        $this->tsena1 = (float) $data['tsena1'] ?? 0;
-        $this->tsena2 = (float) $data['tsena2'] ?? 0;
-        $this->tsena3 = (float) $data['tsena3'] ?? 0;
+        $this->deb_kol_kon = (float)$data['deb_kol_kon'] ?? 0;
+        $this->tsena1 = (float)$data['tsena1'] ?? 0;
+        $this->tsena2 = (float)$data['tsena2'] ?? 0;
+        $this->tsena3 = (float)$data['tsena3'] ?? 0;
     }
 
     public function toArray(): array
